Rename end-point and fix tests
