+++ functions/src/metrics/logFeedback/index.spec.ts
@@ -9,7 +9,7 @@ import {addFeedback} from '../../controllers/feedback';
 jest.mock('../../controllers/feedback');
 
 const router = createMetricsRouter();
-router.use('/feedback', feedbackRouter.routes());
+router.use('/logFeedback', feedbackRouter.routes());
 const mockServer = createMockServer(router.routes(), router.allowedMethods());
 
 afterEach(() => {
@@ -20,11 +20,14 @@ afterAll(() => {
   mockServer.close();
 });
 
-describe('/metrics/feedback', () => {
+describe('/metrics/logFeedback', () => {
   describe('Success', () => {
     it('Accepts incoming feedback', async () => {
-      const response = await request(mockServer).post('/feedback').send({
+      const response = await request(mockServer).post('/logFeedback').send({
         exerciseId: 'some-exercise-id',
+        completed: true,
+        sessionId: 'some-session-id',
+        host: false,
         question: 'Some question?',
         answer: true,
         comment: 'Some comment!',
@@ -33,6 +36,9 @@ describe('/metrics/feedback', () => {
       expect(addFeedback).toHaveBeenCalledTimes(1);
       expect(addFeedback).toHaveBeenCalledWith({
         exerciseId: 'some-exercise-id',
+        completed: true,
+        sessionId: 'some-session-id',
+        host: false,
         question: 'Some question?',
         answer: true,
         comment: 'Some comment!',
@@ -41,9 +47,10 @@ describe('/metrics/feedback', () => {
       expect(response.status).toBe(200);
     });
 
-    it('Accepts undefined comment', async () => {
-      const response = await request(mockServer).post('/feedback').send({
+    it('Accepts undefined sessionId, host and comment', async () => {
+      const response = await request(mockServer).post('/logFeedback').send({
         exerciseId: 'some-exercise-id',
+        completed: true,
         question: 'Some question?',
         answer: true,
       });
@@ -51,6 +58,7 @@ describe('/metrics/feedback', () => {
       expect(addFeedback).toHaveBeenCalledTimes(1);
       expect(addFeedback).toHaveBeenCalledWith({
         exerciseId: 'some-exercise-id',
+        completed: true,
         question: 'Some question?',
         answer: true,
       });
@@ -61,7 +69,20 @@ describe('/metrics/feedback', () => {
 
   describe('Failure', () => {
     it('Requires exerciseId', async () => {
-      const response = await request(mockServer).post('/feedback').send({
+      const response = await request(mockServer).post('/logFeedback').send({
+        completed: true,
+        question: 'Some question?',
+        answer: true,
+      });
+
+      expect(addFeedback).toHaveBeenCalledTimes(0);
+
+      expect(response.status).toBe(500);
+    });
+
+    it('Requires completed', async () => {
+      const response = await request(mockServer).post('/logFeedback').send({
+        exerciseId: 'some-exercise-id',
         question: 'Some question?',
         answer: true,
       });
@@ -72,8 +93,9 @@ describe('/metrics/feedback', () => {
     });
 
     it('Requires question', async () => {
-      const response = await request(mockServer).post('/feedback').send({
+      const response = await request(mockServer).post('/logFeedback').send({
         exerciseId: 'some-exercise-id',
+        completed: true,
         answer: true,
       });
 
@@ -83,8 +105,9 @@ describe('/metrics/feedback', () => {
     });
 
     it('Requires answer', async () => {
-      const response = await request(mockServer).post('/feedback').send({
+      const response = await request(mockServer).post('/logFeedback').send({
         exerciseId: 'some-exercise-id',
+        completed: true,
         question: 'Some question?',
       });
